Port to NDI SDK v4
Most notably this allows the video line stride to be set to 0, in which
case a default value is to be used, and it deprecates selecting NDI
sources by their IP address and because of that we remove the relevant
properties.

Usually the SDK will give an URL instead of an IP address now, so usage
would've been broken anyway.